third_party/WebKit/LayoutTests/web-animations-api/player.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div id='element'></div>

<script>
var element = document.getElementById('element');
test(function() {
    var animation = element.animate([{left: '0px', offset: 0}, {left: '100px', offset: 1}], 10);
    var player = animation.player;
    assert_equals(player.source, animation);
    assert_equals(player.currentTime, 0);
    player.currentTime = 5;
    assert_equals(player.currentTime, 5);
    assert_equals(player.timeLag, -5);

    player.startTime += 1;
    assert_equals(player.currentTime, 4);

    player.pause();
    assert_true(player.paused);
    player.play();
    assert_false(player.paused);

    player.finish();
    assert_equals(player.currentTime, 10);
    assert_true(player.finished);

    player.reverse();
    assert_equals(player.playbackRate, -1);

    player.playbackRate = -2;
    assert_equals(player.playbackRate, -2);

    player.cancel();
    assert_equals(animation.player, null);
}, 'Player sanity tests');

test(function() {
    var player = element.animate([{left: '0px', offset: 0}, {left: '100px', offset: 1}], {iterations: Infinity, duration: 10}).player;
    assert_throws('INVALID_STATE_ERR', function() { player.finish(); });
}, 'Player.finish() raises exception if source content end is infinity');
</script>
